+++ test/regression/input/if.c
@@ -1,13 +1,13 @@
-int64_t if_basic() {
-    int64_t cond = 1;
+int if_basic() {
+    int cond = 1;
     if (cond) {
         cond = cond + 100;
     }
     return cond;
 }
 
-int64_t if_basic_else() {
-    int64_t cond = 1;
+int if_basic_else() {
+    int cond = 1;
     if (cond) {
         cond = cond + 3;
     } else {
@@ -16,25 +16,25 @@ int64_t if_basic_else() {
     return cond;
 }
 
-int64_t if_no_braces() {
-    int64_t cond = 1;
+int if_no_braces() {
+    int cond = 1;
     if (cond)
         return cond;
     else
         return cond;
 }
 
-int64_t if_chain() {
-  int64_t cond = 1;
-  int64_t cond_2 = 2;
+int if_chain() {
+  int cond = 1;
+  int cond_2 = 2;
   if (cond) if (cond_2) return 0;
   return 1;
 }
 
-int64_t if_chain_many() {
-  int64_t cond = 1;
-  int64_t cond_2 = 2;
-  int64_t cond_3 = 0;
+int if_chain_many() {
+  int cond = 1;
+  int cond_2 = 2;
+  int cond_3 = 0;
   if (cond) if (cond_2) {
       if (cond_3) return 0;
       else return 3;
@@ -42,7 +42,7 @@ int64_t if_chain_many() {
   return 1;
 }
 
-int64_t if_else_if() {
+int if_else_if() {
     if (1)
         return 2;
     else if (3)
